+++ IndicatorsTests/Test.ChaikinOscillator.cs
@@ -26,10 +26,10 @@ namespace StockIndicators.Tests
 
             // sample value
             ChaikinOscResult r = results.Where(x => x.Index == 502).FirstOrDefault();
-            Assert.AreEqual((decimal)3439986548.42, Math.Round(r.Adl, 2));
-            Assert.AreEqual((decimal)0.8052, Math.Round(r.MoneyFlowMultiplier, 4));
-            Assert.AreEqual((decimal)118396116.25, Math.Round(r.MoneyFlowVolume, 2));
-            Assert.AreEqual((decimal)-19135200.72, Math.Round((decimal)r.Oscillator, 2));
+            Assert.AreEqual(3439986548.42m, Math.Round(r.Adl, 2));
+            Assert.AreEqual(0.8052m, Math.Round(r.MoneyFlowMultiplier, 4));
+            Assert.AreEqual(118396116.25m, Math.Round(r.MoneyFlowVolume, 2));
+            Assert.AreEqual(-19135200.72m, Math.Round((decimal)r.Oscillator, 2));
         }
