+++ IndicatorsTests/Test.Keltner.cs
@@ -31,16 +31,16 @@ namespace StockIndicators.Tests
 
             // sample value
             KeltnerResult r1 = results.Where(x => x.Index == 502).FirstOrDefault();
-            Assert.AreEqual((decimal)262.1873, Math.Round((decimal)r1.UpperBand, 4));
-            Assert.AreEqual((decimal)249.3519, Math.Round((decimal)r1.Centerline, 4));
-            Assert.AreEqual((decimal)236.5165, Math.Round((decimal)r1.LowerBand, 4));
-            Assert.AreEqual((decimal)0.102950, Math.Round((decimal)r1.Width, 6));
+            Assert.AreEqual(262.1873m, Math.Round((decimal)r1.UpperBand, 4));
+            Assert.AreEqual(249.3519m, Math.Round((decimal)r1.Centerline, 4));
+            Assert.AreEqual(236.5165m, Math.Round((decimal)r1.LowerBand, 4));
+            Assert.AreEqual(0.102950m, Math.Round((decimal)r1.Width, 6));
 
             KeltnerResult r2 = results.Where(x => x.Index == 486).FirstOrDefault();
-            Assert.AreEqual((decimal)275.4260, Math.Round((decimal)r2.UpperBand, 4));
-            Assert.AreEqual((decimal)265.4599, Math.Round((decimal)r2.Centerline, 4));
-            Assert.AreEqual((decimal)255.4938, Math.Round((decimal)r2.LowerBand, 4));
-            Assert.AreEqual((decimal)0.075085, Math.Round((decimal)r2.Width, 6));
+            Assert.AreEqual(275.4260m, Math.Round((decimal)r2.UpperBand, 4));
+            Assert.AreEqual(265.4599m, Math.Round((decimal)r2.Centerline, 4));
+            Assert.AreEqual(255.4938m, Math.Round((decimal)r2.LowerBand, 4));
+            Assert.AreEqual(0.075085m, Math.Round((decimal)r2.Width, 6));
         }
